pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "carevalue-claims-ml"
version = "0.1.0"
description = "ML and analytics pipelines for value based care using a medical claims backbone"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.0",
  "numpy>=1.24",
  "sqlalchemy>=2.0",
  "psycopg2-binary>=2.9",
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "scikit-learn>=1.3",
  "joblib>=1.3",
  "typer>=0.9",
  "rich>=13.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.1.7"
]

[project.scripts]
carevalue-ml = "carevalue_claims_ml.cli:app"

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
